Add file check

+++ _scripts/removeUnusedAttachments.py
@@ -61,4 +61,7 @@ for file in dirList_2:
             break
     # Delete files that are in dirList2 and not in the attachment list
     if deleteFlag is True:
-        pathlib.Path.unlink(file)
+        if pathlib.Path.isFile(file):
+            pathlib.Path.unlink(file)
+        else:
+            continue
